ai_ws/python_playground/requirements.txt
accelerate==1.12.0
aiofile==3.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asgiref==3.11.1
asyncpg==0.31.0
attrs==25.4.0
Authlib==1.6.9
backoff==2.2.1
beartype==0.22.9
beautifulsoup4==4.14.3
bidict==0.23.1
bracex==2.6
bs4==0.0.2
cachetools==7.0.3
caio==0.9.25
casbin==1.43.0
casbin_async_sqlalchemy_adapter==1.17.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.7.0
dataclasses-json==0.6.7
Deprecated==1.3.1
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.124.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastmcp==3.1.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-vision==3.12.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-postgres==0.0.16
langchain-text-splitters==1.1.0
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openapi-pydantic==0.5.1
opencv-python==4.13.0.90
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-flask==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pathable==0.5.0
pdf2image==1.17.0
pdfminer.six==20260107
pgvector==0.3.6
pi_heif==1.1.1
pikepdf==10.2.0
pillow==12.1.0
platformdirs==4.9.4
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
py-key-value-aio==0.4.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycasbin==2.7.1
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.6.0
pypdfium2==5.3.0
pyperclip==1.11.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-socketio==5.16.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==5.0.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.0
sentry-sdk==2.47.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
simpleeval==1.0.3
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
sse-starlette==3.3.2
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uncalled-for==0.2.0
unstructured==0.18.27
unstructured-client==0.42.8
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.7
urllib3==2.6.2
uuid_utils==0.13.0
uvicorn==0.38.0
watchfiles==1.1.1
wcmatch==10.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
